# Conflicts:
#	app.c
#	conf/gfxconf.h
#	debug/mb2016 JLink Flash + Run.launch
#	drivers/drivers.h
#	drivers/eeprom/src/24xx_driver.c
#	drivers/fmc/stm32f4xx_fmc.h
#	drivers/gdisp_lld/gdisp_lld_STM32LTDC.c
#	drivers/gdisp_lld/stm32_dma2d.h
#	main.c
#	modules/ARTracker/ARTracker.c
#	modules/ARTracker/ARTracker.h
#	modules/UDC/udc.c
#	modules/UDC/udc.h
#	modules/UDC/udc_lld.c
#	modules/UDC/udc_lld.h
#	modules/analog/analog.c
#	modules/bubbles/bubbles.c
#	modules/buzzer/buzzer.c
#	modules/buzzer/buzzer.h
#	modules/modules.mk
#	modules/rtt_shell/rtt_shell.c
#	modules/ui/board_testing_subpages/linesensor_test_page.c
#	modules/ui/board_testing_subpages/ps4_testing_page.c
#	modules/ui/menu/menu.c
#	modules/ui/settings_subpages/backlight.c
#	modules/ui/start_robot_page/start_robot_page.c
#	modules/ui/start_robot_page/start_robot_page_gui.c
#	modules/ui/status_bar/status_bar.c
#	modules/ui/template_page/template_page.c
#	modules/ui/tetris/tetris.c
#	modules/ui/ui.mk
#	robot/auto_path.c
#	robot/driving.c
#	robot/driving.h
#	robot/encoder/encoder.c
#	robot/encoder/encoder.h
#	robot/linesensor/linesensor.c
#	robot/linesensor/linesensor.h
#	robot/motor/motor.c
#	robot/motor/motor.h
#	robot/pid.c
#	robot/servo/servo.c
#	robot/servo/servo.h
#	robot/udc_objectlist.h

+++ robot/pid.h
@@ -24,7 +24,7 @@ typedef struct PID_gain
     float kp;
     float ki;
     float kd;
-    int i_term;
+    float i_term;
     int pre_output;
     int pre_input;
 }GAIN;
